backend/cronjob_requirements.txt
scikit-learn
scikit-surprise
pandas 
asyncio
motor
surprise
numpy
dotenv
faiss-cpu
